BUG#: 3826 TITLE: Warning in DynamicConsumer.cpp DESCRIPTION: Add a Uint32 cast to avoid a signed/unsigned comparison compile warning on Linux and Mac platforms.

+++ src/Pegasus/DynListener/DynamicConsumer.cpp
@@ -376,7 +376,7 @@ Boolean DynamicConsumer::isIdle()
     struct timeval timeout = {0,0};
     getIdleTimer(&timeout);
 
-    if ((now.tv_sec - timeout.tv_sec) > (DEFAULT_IDLE_TIMEOUT))  //seconds
+    if (Uint32(now.tv_sec - timeout.tv_sec) > (DEFAULT_IDLE_TIMEOUT)) //seconds
     {
         PEG_METHOD_EXIT();
         return true;
